refactor: change variant category import location
