+++ src/modules/product/dto/create.producto.dto.ts
@@ -12,7 +12,7 @@ import {
   ValidateNested,
 } from 'class-validator';
 import { UpdateVariantDto } from 'src/modules/stock/dto/update-variant.dto';
-import { CreateVariantCategoryDto } from 'src/modules/category/variant/dto/create-variantCategory.dto';
+import { CreateVariantCategoryDto } from 'src/modules/category/variant/dto/create.variant.category.dto';
 
 export class CreateProductDto {
   @IsString()
